refactor: encapsulate tui components
